tests/django_app/pyproject.toml
[project]
name = "pgsqlite-django-test"
version = "0.1.0"
description = "Django REST API test for pgsqlite"
authors = [
    {name = "Test",email = "test@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
